Remove compilation errors from gcc on Solaris

+++ testapp.c
@@ -252,10 +252,10 @@ static enum test_return test_safe_strtol(void) {
 static pid_t start_server(in_port_t *port_out, bool daemon) {
     char environment[80];
     snprintf(environment, sizeof(environment),
-             "MEMCACHED_PORT_FILENAME=/tmp/ports.%u", getpid());
+             "MEMCACHED_PORT_FILENAME=/tmp/ports.%lu", (long)getpid());
     char *filename= environment + strlen("MEMCACHED_PORT_FILENAME=");
     char pid_file[80];
-    snprintf(pid_file, sizeof(pid_file), "/tmp/pid.%u", getpid());
+    snprintf(pid_file, sizeof(pid_file), "/tmp/pid.%lu", (long)getpid());
 
     remove(filename);
     remove(pid_file);
